Refactor schema validation code into multiple files
